Fix Collapsible calculating height all the time
The check in `componentWillReceiveProps` was incorrect, leading to calculations being performed on every prop change.

Since we can now have components that we don’t have the height on first render, we need a check in the Motion height agains the `collapsed` prop, otherwise the Collapsible would show expanded (not opacity) on first render.